source/requirements.snowflake.txt
snowflake-snowpark-python==1.33.0
requests
python-dotenv